Update ArduinoCode
Okay, so my problem here is that each individual capacitor works when you comment out the others, but when all of them are put together, it seems that the console log is very slow and possibly inaccurate. This is my code for the hardware set up. I have also not yet added how we are going to connect each capacitor to processing. I don't know how to send 5 different signals to cause different things to happen in processing. I hope that you know how to make this happen!

The good news is that I practiced using the capacitors with different objects, and you get different values, which should allow us to have different values cause different things to happen in processing, correct? Sorry my code knowledge is very bad! Just started coding in January.